Move LexerRunner to own package
