+++ src/main/java/slp/core/CLI.java
@@ -15,8 +15,8 @@ import slp.core.example.JavaRunner;
 import slp.core.example.NLRunner;
 import slp.core.io.Writer;
 import slp.core.lexing.Lexer;
-import slp.core.lexing.LexerRunner;
 import slp.core.lexing.code.JavaLexer;
+import slp.core.lexing.runners.LexerRunner;
 import slp.core.lexing.simple.CharacterLexer;
 import slp.core.lexing.simple.PunctuationLexer;
 import slp.core.lexing.simple.TokenizedLexer;
